+++ src/Resources/contao/config/config.php
@@ -13,7 +13,8 @@
 
 $GLOBALS['BE_MOD']['content']['topwertungszahlen'] = array
 (
-	'tables'            => array('tl_topwertungszahlen', 'tl_topwertungszahlen_ratings', 'tl_topwertungszahlen_photos')
+	'tables'            => array('tl_topwertungszahlen', 'tl_topwertungszahlen_ratings', 'tl_topwertungszahlen_photos'),
+	'importRating'      => array('Schachbulle\ContaoTopwertungszahlenBundle\Classes\Import', 'run')
 );
 
 /**