+++ src/Resources/contao/languages/de/tl_topwertungszahlen.php
@@ -1,6 +1,24 @@
 <?php
 
 /**
+ * Buttons für globale Operationen
+ */
+$GLOBALS['TL_LANG']['tl_topwertungszahlen']['importRating'] = array('Wertungszahlen aktualisieren', 'Wertungszahlen laden und JSON-Dateien aktualisieren');
+$GLOBALS['TL_LANG']['tl_topwertungszahlen']['importRating_confirm'] = 'Der Dienst läuft einmal täglich als Cronjob. Wollen Sie die aktuellen Wertungszahlen außer der Reihe laden und die JSON-Dateien wirklich aktualisieren?';
+
+/**
+ * Buttons für Operationen
+ */
+$GLOBALS['TL_LANG']['tl_topwertungszahlen']['new'] = array('Neuer Spieler', 'Neuen Spieler anlegen');
+$GLOBALS['TL_LANG']['tl_topwertungszahlen']['edit'] = array("Wertungszahlen des Spielers %s bearbeiten", "Wertungszahlen des Spielers %s bearbeiten");
+$GLOBALS['TL_LANG']['tl_topwertungszahlen']['editHeader'] = array("Spieler %s bearbeiten", "Spieler %s bearbeiten");
+$GLOBALS['TL_LANG']['tl_topwertungszahlen']['editPhoto'] = array("Fotos des Spielers %s bearbeiten", "Fotos des Spielers %s bearbeiten");
+$GLOBALS['TL_LANG']['tl_topwertungszahlen']['copy'] = array("Spieler %s kopieren", "Spieler %s kopieren");
+$GLOBALS['TL_LANG']['tl_topwertungszahlen']['delete'] = array("Spieler %s löschen", "Spieler %s löschen");
+$GLOBALS['TL_LANG']['tl_topwertungszahlen']['toggle'] = array("Spieler %s aktivieren/deaktivieren", "Spieler %s aktivieren/deaktivieren");
+$GLOBALS['TL_LANG']['tl_topwertungszahlen']['show'] = array("Spielerdetails von %s anzeigen", "Details des Spielers %s anzeigen");
+
+/**
  * Backend-Modul: Übersetzungen im Eingabeformular
  */
 $GLOBALS['TL_LANG']['tl_topwertungszahlen']['name_legend'] = 'Personendaten';
@@ -15,14 +33,3 @@ $GLOBALS['TL_LANG']['tl_topwertungszahlen']['dewis_id'] = array('DeWIS-ID', 'ID 
 $GLOBALS['TL_LANG']['tl_topwertungszahlen']['publish_legend'] = 'Veröffentlichung';
 $GLOBALS['TL_LANG']['tl_topwertungszahlen']['published'] = array('Veröffentlicht', 'Spieler aktivieren oder deaktivieren');
 
-/**
- * Buttons für Operationen
- */
-
-$GLOBALS['TL_LANG']['tl_topwertungszahlen']['new'] = array('Neuer Spieler', 'Neuen Spieler anlegen');
-$GLOBALS['TL_LANG']['tl_topwertungszahlen']['edit'] = array('Wertungszahlen des Spielers bearbeiten', 'Wertungszahlen des Spielers %s bearbeiten');
-$GLOBALS['TL_LANG']['tl_topwertungszahlen']['editHeader'] = array('Spieler bearbeiten', 'Spieler %s bearbeiten');
-$GLOBALS['TL_LANG']['tl_topwertungszahlen']['copy'] = array('Spieler kopieren', 'Spieler %s kopieren');
-$GLOBALS['TL_LANG']['tl_topwertungszahlen']['delete'] = array('Spieler löschen', 'Spieler %s löschen');
-$GLOBALS['TL_LANG']['tl_topwertungszahlen']['toggle'] = array('Spieler aktivieren/deaktivieren', 'Spieler %s aktivieren/deaktivieren');
-$GLOBALS['TL_LANG']['tl_topwertungszahlen']['show'] = array('Spielerdetails anzeigen', 'Details des Spielers %s anzeigen');